requirements.txt
argparse==1.4.0
opencv-python==4.10.0.84
pillow==11.0.0
numpy==1.26.4
